post description updating added

+++ src/components/Profile.js
@@ -9,31 +9,66 @@ export default function UserProfile() {
 
     const { usersPosts, fetchUsersPosts } = useContent()
     const [openedPost, setOpenedPost] = useState()
+    const [updatingPost, setUpdatingPost] = useState()
+    const [description, setDescription] = useState()
     const modalRef = useRef()
+    const updateRef = useRef()
 
     useEffect(() => {
         fetchUsersPosts()
     }, [])
 
     const handleDelete = async (docId, img) => {
-        await firestore.collection("posts").doc(docId).delete().then(() => {
-            const fileRef = storage.refFromURL(img)
-            fileRef.delete().then(() => {
-                fetchUsersPosts()
-                Swal.fire(
-                    "Post deleted!",
-                    "",
-                    "success"
-                )
 
-            }).catch((error) => {
-                console.error(error)
-                Swal.fire({
-                    icon: "error",
-                    title: "",
-                    text: "error"
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                await firestore.collection("posts").doc(docId).delete().then(() => {
+                    const fileRef = storage.refFromURL(img)
+                    fileRef.delete().then(() => {
+                        fetchUsersPosts()
+                        Swal.fire(
+                            "Post deleted!",
+                            "",
+                            "success"
+                        )
+                    }).catch((error) => {
+                        console.error(error)
+                        Swal.fire({
+                            icon: "error",
+                            title: "",
+                            text: "error"
+                        })
+                    })
+                }).catch((error) => {
+                    console.error(error)
+                    Swal.fire({
+                        icon: "error",
+                        title: "",
+                        text: "error"
+                    })
                 })
-            })
+            }
+        })
+    }
+
+    const updatePost= async () =>{
+        await firestore.collection("posts").doc(updatingPost.docId).update({description: description}).then(()=>{
+            fetchUsersPosts()
+            updateRef.current.click()
+            Swal.fire(
+                "Post updated!",
+                "",
+                "success"
+            )
         }).catch((error) => {
             console.error(error)
             Swal.fire({
@@ -44,17 +79,49 @@ export default function UserProfile() {
         })
     }
 
-
     const openPost = (e, post) => {
         e.preventDefault()
         setOpenedPost(post)
         modalRef.current.click()
     }
 
+    const editPost = (e, updatingPost) => {
+        e.preventDefault()
+        setUpdatingPost(updatingPost)
+        setDescription(updatingPost.data.description)
+        updateRef.current.click()
+    }
+
     return (
         <>
+            <input type="checkbox" id="updatingPost" className="modal-toggle" ref={updateRef} />
+            <div className="modal">
+                <div className="modal-box">
+                    {updatingPost && (
+                        <>
+                            <img src={updatingPost.data.fileUrl} className="mx-auto" style={{ objectFit: "cover" }} alt={updatingPost.data.description} />
+                            <div className="form-control">
+                                <label className="label">
+                                    <span className="label-text">New Description</span>
+                                </label>
+                                <textarea className="textarea h-24 textarea-bordered textarea-primary" value={description} onChange={e => { setDescription(e.target.value) }} placeholder="New Description"></textarea>
+                            </div>
+                            <div className="modal-action">
+                                <button htmlFor="updatingPost" className="btn btn-primary" onClick={() => {
+                                    updatePost()
+                                }}>Submit</button>
+                                <label htmlFor="updatingPost" className="btn" onClick={e => {
+                                    e.preventDefault()
+                                    updateRef.current.click()
+                                }}>Close</label>
+                            </div>
+                        </>)}
+                </div>
+            </div>
+
+
             <OpenedPost openedPost={openedPost} modalRef={modalRef} />
-           {usersPosts.length?( <h1 className="text-3xl font-semibold text-center my-5">Your Posts</h1>):(<h1 className="text-3xl font-semibold text-center my-5">No Posts yet</h1>)}
+            {usersPosts.length ? (<h1 className="text-3xl font-semibold text-center my-5">Your Posts</h1>) : (<h1 className="text-3xl font-semibold text-center my-5">No Posts yet</h1>)}
             <div className="flex flex-col">
                 {usersPosts.map((post) => {
                     return (
@@ -78,7 +145,7 @@ export default function UserProfile() {
                                     </a>
                                 </li>
                                 <li className="w-1/2">
-                                    <a className="w-full">
+                                    <a className="w-full" onClick={(e) => { editPost(e, post) }}>
                                         <span className="material-icons-outlined mx-auto">
                                             edit
                                         </span>
